# Conflicts:
#	src/main/kotlin/uk/gov/justice/digital/hmpps/arnscoordinatorapi/integrations/assessment/api/StrengthsAndNeedsApi.kt
#	src/main/kotlin/uk/gov/justice/digital/hmpps/arnscoordinatorapi/integrations/plan/api/SentencePlanApi.kt
#	src/main/kotlin/uk/gov/justice/digital/hmpps/arnscoordinatorapi/integrations/plan/api/SentencePlanApiProperties.kt
#	src/main/kotlin/uk/gov/justice/digital/hmpps/arnscoordinatorapi/oasys/OasysCoordinatorService.kt

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/arnscoordinatorapi/oasys/controller/request/OasysRollbackRequest.kt
@@ -8,11 +8,11 @@ import uk.gov.justice.digital.hmpps.arnscoordinatorapi.oasys.entity.OasysUserDet
 data class OasysRollbackRequest(
   @Schema(description = "The SAN Assessment version number that was returned from the Sign Assessment API call.", example = "2")
   @PositiveOrZero
-  val sanVersionNumber: Int?,
+  val sanVersionNumber: Long?,
 
   @Schema(description = "The plan version number that was returned from the Sign Plan API call.", example = "2")
   @PositiveOrZero
-  val sentencePlanVersionNumber: Int?,
+  val sentencePlanVersionNumber: Long?,
 
   @Schema(description = "OASys User Details")
   @Valid
